Bugfix
Got around an annoying buy that happened during user registration, where it would try to access user data in the database before it was created. My solution was to check for the existence of data before accessing it and to log out before returning from the act of registration, thus forcing the user to sign in before the data is accessed. By then it is already created.